refactor(ui): update storybook config path alias
