+++ packages/ui/src/components/Primitive/Image/Image.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react"
 import { Image } from "."
 
 const meta: Meta<typeof Image> = {
-  title: "Components / Image",
+  title: "Components / Primitive / Image",
   component: Image,
   argTypes: {
     src: {
